Conflicts:
	imglib2/examples/src/main/java/interactive/remote/catmaid/CATMAIDRandomAccessibleInterval.java

+++ imglib2/algorithms/core/src/main/java/net/imglib2/algorithm/stats/RealBinMapper.java
@@ -2,10 +2,11 @@
  * #%L
  * ImgLib2: a general-purpose, multidimensional image processing library.
  * %%
- * Copyright (C) 2009 - 2012 Stephan Preibisch, Stephan Saalfeld, Tobias
- * Pietzsch, Albert Cardona, Barry DeZonia, Curtis Rueden, Lee Kamentsky, Larry
- * Lindsey, Johannes Schindelin, Christian Dietz, Grant Harris, Jean-Yves
- * Tinevez, Steffen Jaensch, Mark Longair, Nick Perry, and Jan Funke.
+ * Copyright (C) 2009 - 2013 Stephan Preibisch, Tobias Pietzsch, Barry DeZonia,
+ * Stephan Saalfeld, Albert Cardona, Curtis Rueden, Christian Dietz, Jean-Yves
+ * Tinevez, Johannes Schindelin, Lee Kamentsky, Larry Lindsey, Grant Harris,
+ * Mark Hiner, Aivar Grislis, Martin Horn, Nick Perry, Michael Zinsmaier,
+ * Steffen Jaensch, Jan Funke, Mark Longair, and Dimiter Prodanov.
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -69,7 +70,7 @@ public class RealBinMapper <T extends RealType<T>>
 		
 		//Save a little computation time by calculating these only once.
 		binWidth = (1 + maxBin.getRealDouble() - minBin.getRealDouble()) /
-			((double) numBins);
+			(numBins);
 		halfBinWidth = binWidth / 2;
 		
 		minVal = minBin.getRealDouble();
